# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: QiFlow API
  description: |
    The QiFlow payment gateway API. Integrate once, accept Qi everywhere.

    ## Authentication

    QiFlow uses two authentication mechanisms depending on the endpoint group:

    ### JWT Bearer Token (Dashboard endpoints)
    Issued on login / register. Short-lived access token (15 min) + long-lived refresh token (7 days).
    ```
    Authorization: Bearer <accessToken>
    ```

    ### API Key (Programmatic / v1 endpoints)
    Generated from the dashboard or via `POST /merchants/me/api-keys`.
    ```
    X-API-Key: qf_test_xxxxxxxxxxxxxxxx
    ```

    ## Rate Limiting
    Rate-limit headers are returned on every response:
    | Header | Description |
    |---|---|
    | `RateLimit-Limit` | Max requests per window |
    | `RateLimit-Remaining` | Requests left in current window |
    | `RateLimit-Reset` | Seconds until window resets |

    Auth endpoints have stricter limits (5 req/min for login, 3 req/min for register).
    All other endpoints share a global limit of 500 req/15 min.

    ## Error Format
    All errors follow a consistent shape:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: QiFlow
    url: https://qiflow.xyz
servers:
  - url: https://api.qiflow.xyz
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from the merchant dashboard or `POST /merchants/me/api-keys`

  # ── Reusable Schemas ─────────────────────────────────────────────────────────
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid email or password
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address"]
            password: ["Password must be at least 8 characters"]

    # ── Tokens ───────────────────────────────────────────────────────────────────
    Tokens:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Long-lived JWT (7 days)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    # ── Merchant ─────────────────────────────────────────────────────────────────
    Merchant:
      type: object
      required: [id, email, businessName, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        email:
          type: string
          format: email
          example: merchant@example.com
        businessName:
          type: string
          example: My Qi Store
        walletAddress:
          type: string
          nullable: true
          example: "0x00473a216f2b1d382759e612bf6029fa037e95b2"
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"

    # ── API Key ──────────────────────────────────────────────────────────────────
    ApiKey:
      type: object
      required: [id, name, keyPrefix, lastFour, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        name:
          type: string
          example: Production Key
        keyPrefix:
          type: string
          example: qf_test_
        lastFour:
          type: string
          example: a3f2
        isActive:
          type: boolean
          example: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-01T10:00:00.000Z"
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"
    NewApiKey:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          required: [rawKey, warning]
          properties:
            rawKey:
              type: string
              description: Full API key — only shown once. Store it securely.
              example: qf_test_abcdef1234567890a3f2
            warning:
              type: string
              example: Save this API key now. You will not be able to see it again.

    # ── Payment ──────────────────────────────────────────────────────────────────
...
  responses:
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests, please try again later.

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new merchant
      description: |
        Creates a new merchant account. On success, returns JWT tokens and an
        auto-generated API key. **Store the `rawKey` immediately — it is only
        shown once.**

        Rate limited to **3 requests per minute** per IP.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, businessName]
              properties:
                email:
                  type: string
                  format: email
                  example: merchant@example.com
                password:
                  type: string
                  minLength: 8
                  description: Must contain uppercase, lowercase, and a number
                  example: Password123!
                businessName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: My Qi Store
                walletAddress:
                  type: string
                  description: Optional Quai wallet address for receiving payments
                  example: "0x00473a216f2b1d382759e612bf6029fa037e95b2"
      responses:
        '201':
          description: Merchant registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          merchant:
                            $ref: '#/components/schemas/Merchant'
                          tokens:
                            $ref: '#/components/schemas/Tokens'
                          apiKey:
                            $ref: '#/components/schemas/NewApiKey'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: EMAIL_EXISTS
                  message: A merchant account with this email already exists
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: |
        Authenticates a merchant and returns JWT tokens.

        Rate limited to **5 requests per minute** per IP.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: merchant@example.com
                password:
                  type: string
                  example: Password123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          merchant:
                            $ref: '#/components/schemas/Merchant'
                          tokens:
                            $ref: '#/components/schemas/Tokens'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchanges a valid refresh token for a new access + refresh token pair.
        Use this before the access token expires (15 min TTL) to maintain a session.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tokens:
                            $ref: '#/components/schemas/Tokens'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: Invalid or expired refresh token

  # ── Payments ──────────────────────────────────────────────────────────────────
...